backend/requirements.txt
fastapi>=0.115.0
sqlmodel>=0.0.22
uvicorn[standard]>=0.32.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
pydantic>=2.10.0
pydantic-settings>=2.7.0
psycopg2-binary>=2.9.10
alembic>=1.14.0
python-dateutil>=2.9.0
python-slugify>=8.0.4
requests>=2.32.0
httpx>=0.27.0
aiofiles>=24.1.0
pyjwt>=2.10.0
bcrypt==4.0.1
sqlalchemy>=2.0.36
typing-extensions>=4.12.0
python-dotenv>=1.0.1
slowapi>=0.1.9

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0